refactor: process-errorをcommonに切り出す
