+++ packages/command/processor/src/common/index.ts
@@ -1,0 +1,1 @@
+export * from './process-error';
