+++ packages/command/processor/src/common/process-error.ts
@@ -1,0 +1,21 @@
+abstract class ProcessError extends Error { }
+class ProcessInternalError extends ProcessError {
+  constructor(message: string, cause?: Error) {
+    super(message);
+    this.name = "ProcessError";
+    this.cause = cause;
+  }
+}
+class ProcessNotFoundError extends ProcessError {
+  constructor(message: string, cause?: Error) {
+    super(message);
+    this.name = "ProcessError";
+    this.cause = cause;
+  }
+}
+
+export {
+  ProcessError,
+  ProcessInternalError,
+  ProcessNotFoundError,
+};
